updated isce_stamps_processing.sh and isce_stamps_processing.py

+++ isce_stamps_processing.sh
@@ -1,10 +1,26 @@
 #!/bin/bash
 
+
+# start anaconda
+eval "$(/home/ops/miniconda/bin/conda shell.bash hook)"
+
+
+# save work directory
+WORK_DIR=$(pwd)
+
+
 # copy packages from isce_stamps_processing repo 
-cp -rf $HOME/isce_stamps_processing/pkgs/* $HOME/verdi/pkgs/
+if [[ -e $HOME/verdi/pkgs ]]
+then
+    cp -rf $HOME/isce_stamps_processing/pkgs/* $HOME/verdi/pkgs/
+else
+    mkdir $HOME/verdi/pkgs
+    cp -rf $HOME/isce_stamps_processing/pkgs/* $HOME/verdi/pkgs/
+fi
+
 
 # source paths for StaMPS processing
-source $HOME/isce_stamps_processing/pkgs/StaMPS-master/StaMPS_CONFIG.bash
+. $HOME/isce_stamps_processing/pkgs/StaMPS-master/StaMPS_CONFIG.bash
 
 if [[ -e $TRIANGLE_BIN/triangle ]]
 then
@@ -14,6 +30,7 @@ else
     make
     mkdir bin
     cp triangle bin/
+    cd $WORK_DIR
 fi
 
 if [[ -e $SNAPHU_BIN/snaphu ]]
@@ -22,7 +39,8 @@ then
 else
     cd $HOME/verdi/pkgs/snaphu-v2.0.0/src
     make
+    cd $WORK_DIR
 fi
 
 # run StaMPS processing
-python $HOME/isce_stamps_processing/isce_stamps_processing.py
+python $HOME/isce_stamps_processing_dev/isce_stamps_processing.py
